rename fzf to more generic name
